+++ docs/CHANGELOG.md
@@ -1,4 +1,6 @@
-# v2.0
+# Changelogs
+
+## v2.0
 - Implemented dependency injection and container for better modularity.
 - Removed unused projects, classes, and anti-pattern implementations.
 - Refactored service methods to use MediatR for improved maintainability.
@@ -19,3 +21,5 @@
 - Implemented SupportOrHigher, ModeratorOrHigher, and other hierarchical policies for flexible role-based access control.
 - Added a simple account view dialog for users to manage their account details.
 - Improved the authorization system to dynamically evaluate permissions from the database while still supporting attribute-based authorization.
+- Fixed a bug when selecting ticket type, message update was not being done correctly
+- Fully fixed avarage response time calculation, only calculates the first responses between mods and user.